requirements.txt
aiohappyeyeballs==2.6.1
aiohttp==3.13.3
aiosignal==1.4.0
anyio==4.12.1
attrs==25.4.0
certifi==2026.1.4
click==8.3.1
contourpy==1.3.3
cycler==0.12.1
dwpose==1.0.1
filelock==3.20.0
flatbuffers==25.12.19
fonttools==4.61.1
frozenlist==1.8.0
fsspec==2025.12.0
h11==0.16.0
hf-xet==1.2.0
httpcore==1.0.9
httpx==0.28.1
huggingface_hub==1.3.4
idna==3.11
ImageIO==2.37.2
Jinja2==3.1.6
kiwisolver==1.4.9
lazy_loader==0.4
lightning-utilities==0.15.2
lpips==0.1.4
MarkupSafe==3.0.3
matplotlib==3.10.8
mpmath==1.3.0
multidict==6.7.0
networkx==3.6.1
numpy==2.4.2
nvidia-cublas-cu12==12.8.4.1
nvidia-cuda-cupti-cu12==12.8.90
nvidia-cuda-nvrtc-cu12==12.8.93
nvidia-cuda-runtime-cu12==12.8.90
nvidia-cudnn-cu12==9.10.2.21
nvidia-cufft-cu12==11.3.3.83
nvidia-cufile-cu12==1.13.1.3
nvidia-curand-cu12==10.3.9.90
nvidia-cusolver-cu12==11.7.3.90
nvidia-cusparse-cu12==12.5.8.93
nvidia-cusparselt-cu12==0.7.1
nvidia-nccl-cu12==2.27.5
nvidia-nvjitlink-cu12==12.8.93
nvidia-nvshmem-cu12==3.3.20
nvidia-nvtx-cu12==12.8.90
onnxruntime-gpu==1.24.2
opencv-contrib-python==4.13.0.92
opencv-python==4.12.0.88
openpose-pytorch @ git+https://github.com/prasunroy/openpose-pytorch.git@6a03d76cc52b4eba93762dce1687681fad73b0ee
packaging==26.0
pillow==12.0.0
propcache==0.4.1
protobuf==6.33.5
pyparsing==3.3.2
python-dateutil==2.9.0.post0
pytorch-lightning==2.6.0
PyYAML==6.0.3
scikit-image==0.26.0
scipy==1.17.0
setuptools==80.9.0
shellingham==1.5.4
six==1.17.0
sympy==1.14.0
tifffile==2026.1.14
torch==2.9.1
torch-lr-finder==0.2.2
torchmetrics==1.8.2
torchvision==0.24.1
tqdm==4.67.1
triton==3.5.1
typer-slim==0.21.1
typing_extensions==4.15.0
yarl==1.22.0
